Breaking everything!
We decided to replace getLastCommitBy with a boolean hasCommittedSince which is much simpler. We're half way through the implementaion, so some of the code which was using the original method will need to be updated.

Look at the TODOs and build errors when you run `npm run test` for clues!

Co-authored-by: Blaise Pabon <blaise@gmail.com>
Co-authored-by: Kate Dames <kate.funficient@gmail.com>
Co-authored-by: Olle Jonsson <olle.jonsson@gmail.com>