Re-design configuration for nulls
We also started using prettier in our editor so some files got tidied
up.

We re-designed the null config by thinking about it more from the point
of view of the client code (i.e. the Cucumber scenarios) that would be
using it.

Although we didn't try to use it from the Cucumber scenarios yet, we
just worked with a test that calls retireInacticeContributors directly.

Next we'll try and use it from the Cucumber tests and see how that
impacts things.

Co-authored-by: James Shore <jshore@jamesshore.com>