Create an ADR for containerized npm
TL;DR: I don't use `npm` or TypeScript and other new contributors have arrived with no `npm` environment set up. It would make my life easier to have a container definition so that the tests and builds of the project could run in a container with minimal changes to a contributor's machine.